backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic[email]==2.5.0
python-multipart==0.0.6
redis==5.0.1
websockets==10.4
pytest==7.4.3
pytest-asyncio==0.21.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4